bp - split billing tests in two dinstict test suites
